init core app: refactor email, wordcloud utils
